create lableControler

+++ laravel/app/Http/Controllers/StatusController.php
@@ -1,40 +1,68 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Status;
+use Illuminate\Http\Request;
+
 class StatusController extends Controller
 {
+
     public function index()
     {
         $statuses = Status::all();
-        return view('status', compact('statuses'));
+        return view('status.index', compact('statuses'));
     }
 
     public function create()
     {
-        $status = [
-            'name' => 'exampl 3'
-        ];
-        
-        Status::create($status);
-        dd('create');
+        $status = new Status();
+        return view('status.create', compact('status'));
     }
 
-    public function update()
+    public function store(Request $request)
     {
-        $status = Status::find(1);
-        $status->update([
-            'name' => 'kek'
+        $data = $this->validate($request, [
+            'name' => 'required|unique:statuses',
         ]);
 
-        dd('update');
+        $status = new Status();
+        // Заполнение статьи данными из формы
+        $status->fill($data);
+        // При ошибках сохранения возникнет исключение
+        $status->save();
+
+        // Редирект на указанный маршрут
+        return redirect()->route('status.index');
     }
 
-    public function delete()
+    public function edit($id)
     {
-        $status = Status::find(1);
-        $status->delete();
+        $status = Status::findOrFail($id);
+        return view('status.edit', compact('status'));
+    }
 
-        dd('delete');
+    public function update(Request $request, $id)
+    {
+        $status = Status::findOrFail($id);
+        $data = $this->validate($request, [
+            // У обновления немного измененная валидация. В проверку уникальности добавляется название поля и id текущего объекта
+            // Если этого не сделать, Laravel будет ругаться на то что имя уже существует
+            'name' => 'required|unique:statuses,name,' . $status->id,
+        ]);
+
+        $status->fill($data);
+        $status->save();
+        return redirect()
+            ->route('status.index');
+    }
+
+    public function destroy($id)
+    {
+        $status = status::find($id);
+    if ($status) {
+      $status->delete();
+    }
+    return redirect()->route('status.index');
     }
 }
